bookr/requirements.txt
asgiref==3.2.10
Django==3.1
pytz==2023.3
sqlparse==0.4.4

djangorestframework~=3.14.0